fastapi/requirements.txt
requests~=2.31.0
python-dotenv~=1.0.0
praw~=7.7.1
fastapi[all]==0.110.0
Faker==24.2.0

pydantic==2.6.4
prawcore~=2.4.0
confluent-kafka==2.3.0
fastavro==1.9.4
uvicorn==0.29.0
